Fix to 'resume' function
The 'resume' function seemed to be looking for an incorrect playerState value. It was looking for a value of 1 (playing) in order to issue playVideo. This would not work in the scenario where the player was already paused (e.g. due to a previous call to the 'suspend' function. The 'resume' function has been corrected to look for a playerState value of 2 (paused) in order to playVideo. This change ensures that playback is paused when the module is hidden and then played when the module is shown.